init strategist mod; split away bridge logic into new mod
