+++ e2e/Cargo.toml
@@ -72,3 +72,4 @@ bech32                               = "0.11.0"
 bollard                              = "0.18.1"
 futures-util                         = "0.3"
 toml                                 = "0.8.19"
+async-trait                          = "0.1.77"
