+++ e2e/examples/eth_vault/strategist/mod.rs
@@ -1,0 +1,2 @@
+mod bridge;
+pub mod client;
